Cleanup: refactor to use single `JsonEquals` method.
