+++ test/Serilog.Sinks.Elasticsearch.Tests/Templating/Sendsv7TemplateTests.cs
@@ -36,9 +36,7 @@ namespace Serilog.Sinks.Elasticsearch.Tests.Templating
         [Fact]
         public void ShouldRegisterTheCorrectTemplateOnRegistration()
         {
-
-            var method = typeof(Sendsv7TemplateTests).GetMethod(nameof(ShouldRegisterTheCorrectTemplateOnRegistration));
-            JsonEquals(_templatePut.Item2, method, "template_v7.json");
+            JsonEquals(_templatePut.Item2, "template_v7.json");
         }
 
         [Fact]
@@ -46,22 +44,6 @@ namespace Serilog.Sinks.Elasticsearch.Tests.Templating
         {
             var uri = _templatePut.Item1;
             uri.AbsolutePath.Should().Be("/_template/serilog-events-template");
-        }
-
-        protected void JsonEquals(string json, MethodBase method, string fileName = null)
-        {
-#if DOTNETCORE
-            var assembly = typeof(Sendsv7TemplateTests).GetTypeInfo().Assembly;
-#else
-            var assembly = Assembly.GetExecutingAssembly();
-#endif
-            var expected = TestDataHelper.ReadEmbeddedResource(assembly, fileName ?? "template.json");
-
-            var nJson = JObject.Parse(json);
-            var nOtherJson = JObject.Parse(expected);
-            var equals = JToken.DeepEquals(nJson, nOtherJson);
-            if (equals) return;
-            expected.Should().BeEquivalentTo(json);
         }
     }
 }
