+++ test/Serilog.Sinks.Elasticsearch.Tests/Templating/SetZeroReplicasInTemplateTests.cs
@@ -35,8 +35,7 @@ namespace Serilog.Sinks.Elasticsearch.Tests.Templating
         [Fact]
         public void ShouldRegisterTheCorrectTemplateOnRegistration()
         {
-            var method = typeof(SendsTemplateTests).GetMethod(nameof(ShouldRegisterTheCorrectTemplateOnRegistration));
-            JsonEquals(_templatePut.Item2, method, "template");
+            JsonEquals(_templatePut.Item2, "template_0replicas.json");
         }
 
         [Fact]
@@ -44,22 +43,6 @@ namespace Serilog.Sinks.Elasticsearch.Tests.Templating
         {
             var uri = _templatePut.Item1;
             uri.AbsolutePath.Should().Be("/_template/serilog-events-template");
-        }
-
-        protected void JsonEquals(string json, MethodBase method, string fileName = null)
-        {
-#if DOTNETCORE
-            var assembly = typeof(SendsTemplateTests).GetTypeInfo().Assembly;
-#else
-            var assembly = Assembly.GetExecutingAssembly();
-#endif
-            var expected = TestDataHelper.ReadEmbeddedResource(assembly, "template_0replicas.json");
-
-            var nJson = JObject.Parse(json);
-            var nOtherJson = JObject.Parse(expected);
-            var equals = JToken.DeepEquals(nJson, nOtherJson);
-            if (equals) return;
-            expected.Should().BeEquivalentTo(json);
         }
     }
 }
